deploy/agent/requirements.txt
strands-agents
strands-agents-tools
mcp
fastapi
uvicorn[standard]
pydantic
sse-starlette
